backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langchain
transformers
torch
torchvision
pillow
requests
aiohttp
SpeechRecognition
gtts
pandas
numpy
scikit-learn
firebase-admin
pydantic
python-jose[cryptography]
passlib[bcrypt]
beautifulsoup4
crawl4ai
playwright
selenium
requests-html
pyppeteer
aiohttp
asyncio-mqtt
pyyaml
lxml
chromadb
tiktoken
chromadb[sqlite]
tiktoken 

langchain 
langchain-openai
langgraph
